chore: move FeaturedEventCard to cards
