+++ website-frontend/src/lib/components/cards/FeaturedEventCard.svelte
