+++ website-frontend/src/routes/events/[slug]/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	/** @type {import('./$types').PageData} */
 	import DetailsBanner from '$lib/components/events/DetailsBanner.svelte';
-	import FeaturedEventCard from '$lib/components/events/FeaturedEventCard.svelte';
+	import FeaturedEventCard from '$lib/components/cards/FeaturedEventCard.svelte';
 	import FlexibleContent from '$lib/components/flexible_content/FlexibleContent.svelte';
 
 	export let data;
